requirements.txt
streamlit
pandas
numpy
altair
python-dateutil
matplotlib
pillow
